backend/pyproject.toml
[tool.poetry]
name = "news-portal-backend"
version = "0.1.0"
description = ""
authors = ["scxrchy <scxrchybeats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
ruff = "^0.6.9"
django-split-settings = "^1.3.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"
django-cors-headers = "^4.4.0"
gunicorn = "^23.0.0"
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"
django-jazzmin = "^3.0.1"
daphne = "^4.1.2"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
requests = "^2.32.3"
aiohttp = "^3.10.9"
retrying = "^1.3.4"
ijson = "^3.3.0"
django-filter = "^24.3"
jsonschema = "^4.23.0"
djangorestframework-simplejwt = "^5.3.1"
pandas = "^2.2.3"
flower = "^2.0.1"
python-json-logger = "^2.0.7"
faker = "^30.3.0"
twisted = {extras = ["tls"], version = "^24.7.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
